refactor(api): extract runs routes to src/api/runs_routes.py [Patch 5/9]
Move 4 runs routes (code, pine, detail, list) and 4 exclusive helpers
from api_server.py into a self-contained route module following the
PR #375 pattern.

- runs_routes.py defines _load_json_file, _load_csv_to_dict,
  _build_response_from_run_dir, _run_response_payload locally
- register_runs_routes() resolves auth deps and shared Pydantic models
  (RunResponse, RunInfo, BacktestMetrics, etc.) via sys.modules fallback
- Late-access closures for RUNS_DIR (monkeypatch target) and
  _validate_path_param (shared path safety function)
- Re-exports: _build_response_from_run_dir, _load_json_file,
  _load_csv_to_dict for test compatibility (test_run_card.py calls
  _build_response_from_run_dir 5x directly)
- api_server.py: -357 lines (3155 -> 2798)

Shared symbols that stay in api_server.py: RUNS_DIR,
_validate_path_param (_SAFE_PATH_PARAM_RE), all Pydantic models
(RunResponse, RunInfo, Artifact, RAGSelection, BacktestMetrics),
build_run_analysis and load_run_context (late imports from src.ui_services).

AI-Model: idealab/Qwen3.7-Max-DogFooding
AI-Contributed/Feature: 812/812
AI-Contributed/UT: 0/0
